Split core version tests into dedicated module
